revise build readme and streamline main readme (#28)
Mostly updates and organizes brick_build documentation, but there was some stuff in the main README that didn't need to be there. The biggest change is the splitting up of provider generators and domain generators.

Additionally, the advised API considerations are separated. The code generation documentation focuses on code generation.